# docs/api.yaml
openapi: 3.0.3
info:
  title: AviAI API
  version: 0.3.0
  description: >
    AviAI 后端接口文档（OpenAPI 3.0）。基础路径为 /api/v1，
    统一响应结构为 code/message/data，鉴权方式为 Bearer Token（JWT）。
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: 本地开发环境
...
paths:
...
  /birds/recognize:
    post:
      tags: [Birds]
      summary: 上传图片识别
      description: 上传鸟类图片，触发识别流程并生成一条识别记录。
      operationId: recognizeBird
      x-idempotent: false
      security:
        - BearerAuth: []
      requestBody:
        required: true
        description: 图片文件（jpg/png，建议 <= 5MB）
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: 待识别图片
      responses:
        "201":
          description: 识别成功并创建记录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecognizeResponse"
              examples:
                success:
                  value:
                    code: 0
                    message: ok
                    data:
                      recordId: 101
                      birdName: 白鹭
                      confidence: 0.9342
                      imageUrl: /uploads/20260309_xxx.jpg
                      createdAt: "2026-03-09T10:35:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /birds/records:
    get:
      tags: [Birds]
      summary: 查询识别记录
      description: 分页查询当前用户的识别历史记录。
      operationId: listBirdRecords
      x-idempotent: true
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BirdRecordPageResponse"
              examples:
                success:
                  value:
                    code: 0
                    message: ok
                    data:
                      list:
                        - recordId: 101
                          birdName: 白鹭
                          confidence: 0.9342
                          imageUrl: /uploads/20260309_xxx.jpg
                          createdAt: "2026-03-09T10:35:00Z"
                      total: 1
                      page: 1
                      pageSize: 10
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 身份验证
        使用格式：Authorization: Bearer <token>
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从 1 开始。
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: 每页数量。
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
...
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 1001
            message: 参数错误
            data: null
    Unauthorized:
      description: 未登录或令牌无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 1002
            message: 未登录或 Token 无效
            data: null
...
    InternalError:
      description: 服务内部异常
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 1005
            message: 服务内部错误
            data: null
  schemas:
...
    ErrorResponse:
      type: object
      description: 统一错误响应结构。
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 错误码。
          example: 1001
        message:
          type: string
          description: 错误描述。
          example: 参数错误
        data:
          type: object
          nullable: true
          description: 错误附加信息，默认 null。
          additionalProperties: true
          example: null
...
    RecognizeResponseData:
      type: object
      description: 图片识别结果。
      required: [recordId, birdName, confidence, imageUrl, createdAt]
      properties:
        recordId:
          type: integer
          description: 识别记录 ID。
          example: 101
        birdName:
          type: string
          description: 识别出的鸟类名称。
          example: 白鹭
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: 识别置信度（0~1）。
          example: 0.9342
        imageUrl:
          type: string
          description: 图片存储地址。
          example: /uploads/20260309_xxx.jpg
        createdAt:
          type: string
          format: date-time
          description: 识别时间。
          example: "2026-03-09T10:35:00Z"
    RecognizeResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: "#/components/schemas/RecognizeResponseData"
    BirdRecordItem:
      type: object
      description: 识别记录项。
      required: [recordId, birdName, confidence, imageUrl, createdAt]
      properties:
        recordId:
          type: integer
          description: 记录 ID。
          example: 101
        birdName:
          type: string
          description: 鸟类名称。
          example: 白鹭
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: 置信度。
          example: 0.9342
        imageUrl:
          type: string
          description: 图片 URL。
          example: /uploads/20260309_xxx.jpg
        createdAt:
          type: string
          format: date-time
          description: 识别时间。
          example: "2026-03-09T10:35:00Z"
    BirdRecordPageData:
      type: object
      description: 识别记录分页数据。
      required: [list, total, page, pageSize]
      properties:
        list:
          type: array
          description: 记录列表。
          items:
            $ref: "#/components/schemas/BirdRecordItem"
        total:
          type: integer
          description: 总记录数。
          example: 1
        page:
          type: integer
          description: 当前页码。
          example: 1
        pageSize:
          type: integer
          description: 每页条数。
          example: 10
    BirdRecordPageResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          $ref: "#/components/schemas/BirdRecordPageData"
...